+++ module-sys/Service/Service.cpp
@@ -57,7 +57,7 @@ namespace sys
 
             auto ret = msg->Execute(this);
             if (ret == nullptr) {
-                LOG_FATAL("NO MESSAGE from: %s msg type: %d", msg->sender.c_str(), msg->type);
+                LOG_FATAL("NO MESSAGE from: %s msg type: %d", msg->sender.c_str(), static_cast<int>(msg->type));
                 ret = std::make_shared<DataMessage>(MessageType::MessageTypeUninitialized);
             }
